requirements.txt
Jinja2==3.1.*
pandas==2.2.*
openpyxl==3.1.*
python-dotenv==1.0.*
pathlib==1.*
